+++ modules/emulation/retroarch.nix
@@ -29,9 +29,9 @@ let
     input_exit_emulator_btn = "10"
     input_save_state_btn = "6"
     input_load_state_btn = "5"
-    input_reset_btn = "0"
+    input_reset_btn = "1"
     input_fps_toggle_btn = "3"
-    input_screenshot_btn = "1"
+    input_screenshot_btn = "0"
     input_hold_fast_forward_btn = "8"
     config_save_on_exit = "false"
     log_verbosity = "true"
@@ -44,7 +44,22 @@ let
     threaded_video = "false"
     auto_remaps_enable = "true"
     remap_directory = "${cfg.configRoot}/retroarch/remaps"
-    core_options_path = "${cfg.configRoot}/retroarch/core-options/default.opt"
+    cheevos_enable = "true"
+    cheevos_hardcore_mode_enable = "false"
+    cheevos_verbose_enable = "true"
+    cheevos_start_active = "true"
+    cheevos_auto_screenshot = "true"
+    cheevos_badges_enable = "true"
+    cheevos_challenge_indicators = "true"
+    cheevos_richpresence_enable = "true"
+    cheevos_visibility_account = "true"
+    cheevos_visibility_unlock = "true"
+    cheevos_visibility_mastery = "true"
+    cheevos_visibility_lboard_start = "true"
+    cheevos_visibility_lboard_submit = "true"
+    cheevos_visibility_lboard_trackers = "false"
+    cheevos_unlock_sound_enable = "true"
+    cheevos_test_unofficial = "false"
   '';
 
   globalShaderPreset = pkgs.writeText "emulation-global.slangp" ''
@@ -52,14 +67,14 @@ let
   '';
 
   coreOptions = {
-    "retroarch-fbneo.opt" = ''
+    "FinalBurn Neo" = ''
       fbneo-allow-patched-romsets = "enabled"
     '';
-    "retroarch-fceumm.opt" = ''
+    "FCEUmm" = ''
       fceumm_region = "Auto"
       fceumm_ramstate = "enabled"
     '';
-    "retroarch-beetle-supergrafx.opt" = ''
+    "Beetle SuperGrafx" = ''
       sgx_cdimagecache = "enabled"
       sgx_cdbios = "System Card 3"
       sgx_default_joypad_type_p1 = "6 Buttons"
@@ -69,126 +84,57 @@ let
       sgx_default_joypad_type_p5 = "6 Buttons"
       sgx_multitap = "enabled"
     '';
-    "retroarch-beetle-psx-hw.opt" = ''
+    "Beetle PCE Fast" = ''
+      pce_fast_cdimagecache = "enabled"
+      pce_fast_cdbios = "System Card 3"
+      pce_fast_default_joypad_type_p1 = "6 Buttons"
+      pce_fast_default_joypad_type_p2 = "6 Buttons"
+      pce_fast_default_joypad_type_p3 = "6 Buttons"
+      pce_fast_default_joypad_type_p4 = "6 Buttons"
+      pce_fast_default_joypad_type_p5 = "6 Buttons"
+    '';
+    "Beetle PSX HW" = ''
       beetle_psx_hw_renderer = "hardware_vk"
       beetle_psx_hw_pgxp_mode = "memory + CPU"
       beetle_psx_hw_internal_resolution = "4x"
     '';
-    "retroarch-beetle-saturn.opt" = ''
+    "Beetle Saturn" = ''
       beetle_saturn_virtuagun_crosshair = "Cross"
     '';
-    "retroarch-mupen64plus.opt" = ''
-      mupen64plus-rdp-plugin = "parallel"
+    "Mupen64Plus-Next" = ''
+      mupen64plus-rdp-plugin = "gliden64"
       mupen64plus-cpucore = "dynamic_recompiler"
+      mupen64plus-rsp-plugin = "hle"
+      mupen64plus-EnableNativeResFactor = "3"
+      mupen64plus-aspect = "4:3"
+      mupen64plus-43screensize = "960x720"
+      mupen64plus-EnableFBEmulation = "True"
+      mupen64plus-EnableCopyColorToRDRAM = "Async"
+      mupen64plus-EnableCopyDepthToRDRAM = "Software"
+      mupen64plus-BilinearMode = "3point"
+      mupen64plus-MultiSampling = "0"
+      mupen64plus-FXAA = "0"
+      mupen64plus-EnableShadersStorage = "True"
+      mupen64plus-CountPerOp = "0"
+      mupen64plus-ForceDisableExtraMem = "False"
+      mupen64plus-pak1 = "rumble"
+      mupen64plus-pak2 = "rumble"
+      mupen64plus-pak3 = "rumble"
+      mupen64plus-pak4 = "rumble"
     '';
-    "retroarch-parallel-n64.opt" = ''
-      parallel-n64-gfxplugin = "parallel"
-    '';
-    "retroarch-melonds.opt" = ''
+    "melonDS" = ''
       melonds_boot_directly = "enabled"
     '';
-    "retroarch-ppsspp.opt" = ''
+    "PPSSPP" = ''
       ppsspp_internal_resolution = "4"
     '';
-    "retroarch-pcsx2.opt" = ''
+    "LRPS2" = ''
       pcsx2_renderer = "Vulkan"
       pcsx2_upscale_multiplier = "3"
     '';
-    "retroarch-flycast.opt" = ''
+    "Flycast" = ''
       flycast_renderer = "vulkan"
       reicast_threaded_rendering = "disabled"
-    '';
-  };
-
-  profiles = {
-    "nnedi3-clean.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/edge-smoothing/nnedi3/nnedi3-nns16-2x-rgb.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "nnedi3-quality.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/edge-smoothing/nnedi3/nnedi3-nns32-4x-rgb.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "nnedi3-balanced.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/edge-smoothing/nnedi3/nnedi3-nns32-2x-rgb-nns32-4x-luma.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "nnedi3-fast.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/edge-smoothing/nnedi3/nnedi3-nns16-2x-rgb.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "sharp-bilinear-prescale.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/pixel-art-scaling/sharp-bilinear-2x-prescale.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "sharp-bilinear-simple.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/pixel-art-scaling/sharp-bilinear-simple.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "pixel-aa-fast.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/pixel-art-scaling/pixel_aa_fast.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "scalefx-aa-fast.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/presets/scalefx-plus-smoothing/scalefx-aa-fast.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "xbrz-freescale.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/edge-smoothing/xbrz/xbrz-freescale-multipass.slangp"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "megabezel-auto.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/bezel/Mega_Bezel/Presets/Base_CRT_Presets_DREZ/MBZ__3__STD__GDV__DREZ-VIEWPORT.slangp"
-    '';
-    "megabezel-standard.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/bezel/Mega_Bezel/Presets/Base_CRT_Presets_DREZ/MBZ__3__STD__GDV__DREZ-VIEWPORT.slangp"
-    '';
-    "megabezel-potato.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/bezel/Mega_Bezel/Presets/Base_CRT_Presets_DREZ/MBZ__5__POTATO__GDV__DREZ-VIEWPORT.slangp"
-    '';
-    "megabezel-passthrough.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/bezel/Mega_Bezel/Presets/Base_CRT_Presets_DREZ/MBZ__3__STD__PASSTHROUGH__DREZ-VIEWPORT.slangp"
-    '';
-    "sharp-clean.cfg" = ''
-      video_shader_enable = "true"
-      video_shader = "${cfg.configRoot}/retroarch/shaders/shaders_slang/pixel-art-scaling/sharp-bilinear-2x-prescale.slangp"
-    '';
-    "no-shader.cfg" = ''
-      video_shader_enable = "false"
-      video_smooth = "false"
-      video_scale_integer = "false"
-    '';
-    "integer-raw.cfg" = ''
-      video_shader_enable = "false"
-      video_scale_integer = "true"
-      video_smooth = "false"
-    '';
-    "performance.cfg" = ''
-      video_shader_enable = "false"
-      video_smooth = "false"
-      threaded_video = "true"
     '';
   };
 
@@ -201,6 +147,10 @@ let
     input_y_btn = "3"
     input_select_btn = "8"
     input_start_btn = "9"
+    input_up_btn = "h0up"
+    input_down_btn = "h0down"
+    input_left_btn = "h0left"
+    input_right_btn = "h0right"
     input_a_btn = "0"
     input_x_btn = "2"
     input_l_btn = "4"
@@ -209,6 +159,14 @@ let
     input_r2_axis = "+5"
     input_l3_btn = "10"
     input_r3_btn = "11"
+    input_enable_hotkey_btn = "8"
+    input_menu_toggle_btn = "2"
+    input_save_state_btn = "5"
+    input_load_state_btn = "4"
+    input_reset_btn = "1"
+    input_fps_toggle_btn = "3"
+    input_screenshot_btn = "0"
+    input_hold_fast_forward_axis = "+5"
     input_l_x_plus_axis = "+0"
     input_l_x_minus_axis = "-0"
     input_l_y_plus_axis = "+1"
@@ -217,6 +175,22 @@ let
     input_r_x_minus_axis = "-3"
     input_r_y_plus_axis = "+4"
     input_r_y_minus_axis = "-4"
+    input_b_btn_label = "B"
+    input_y_btn_label = "Y"
+    input_select_btn_label = "Minus"
+    input_start_btn_label = "Plus"
+    input_up_btn_label = "D-Pad Up"
+    input_down_btn_label = "D-Pad Down"
+    input_left_btn_label = "D-Pad Left"
+    input_right_btn_label = "D-Pad Right"
+    input_a_btn_label = "A"
+    input_x_btn_label = "X"
+    input_l_btn_label = "L"
+    input_r_btn_label = "R"
+    input_l2_btn_label = "ZL"
+    input_r2_btn_label = "ZR"
+    input_l3_btn_label = "Left Stick Press"
+    input_r3_btn_label = "Right Stick Press"
   '';
 
   autoconfigSwitchPro = pkgs.writeText "Nintendo Switch Pro Controller.cfg" ''
@@ -242,6 +216,14 @@ let
     input_r2_btn = "8"
     input_l3_btn = "12"
     input_r3_btn = "13"
+    input_enable_hotkey_btn = "9"
+    input_menu_toggle_btn = "2"
+    input_save_state_btn = "6"
+    input_load_state_btn = "5"
+    input_reset_btn = "1"
+    input_fps_toggle_btn = "3"
+    input_screenshot_btn = "0"
+    input_hold_fast_forward_btn = "8"
     input_l_x_plus_axis = "+0"
     input_l_x_minus_axis = "-0"
     input_l_y_plus_axis = "+1"
@@ -268,60 +250,10 @@ let
     input_r3_btn_label = "Right Stick Press"
   '';
 
-  systemShaderDefaults = {
-    fbneo = "nnedi3-clean.cfg";
-    pcengine = "nnedi3-clean.cfg";
-    pcenginecd = "nnedi3-clean.cfg";
-    gb = "nnedi3-clean.cfg";
-    gbc = "nnedi3-clean.cfg";
-    gba = "nnedi3-clean.cfg";
-    n64 = "no-shader";
-    nds = "nnedi3-clean.cfg";
-    nes = "nnedi3-clean.cfg";
-    snes = "nnedi3-clean.cfg";
-    virtualboy = "sharp-bilinear-prescale";
-    neogeocd = "nnedi3-clean.cfg";
-    ngpc = "nnedi3-clean.cfg";
-    dreamcast = "no-shader";
-    gamegear = "nnedi3-clean.cfg";
-    genesis = "nnedi3-clean.cfg";
-    mastersystem = "nnedi3-clean.cfg";
-    saturn = "sharp-bilinear-simple";
-    segacd = "nnedi3-clean.cfg";
-    psx = "sharp-bilinear-simple";
-  };
-
-  shaderPolicy = pkgs.writeText "emulation-shader-policy.json" (builtins.toJSON {
-    default = "nnedi3-clean";
-    fallback = "sharp-bilinear-prescale";
-    profiles = builtins.attrNames profiles;
-    systemDefaults = systemShaderDefaults;
-    cleanScaling = {
-      quality = "nnedi3-quality";
-      balanced = "nnedi3-balanced";
-      fast = "nnedi3-fast";
-      sharp = "sharp-bilinear-prescale";
-      raw = "integer-raw";
-    };
-    megabezel = {
-      requiredPath = "shaders_slang/bezel/Mega_Bezel";
-      highQuality = "megabezel-standard";
-      lowCost = "megabezel-potato";
-      passthrough = "megabezel-passthrough";
-    };
-  });
-
-  profileFiles = lib.mapAttrsToList
-    (name: text: {
-      inherit name;
-      file = pkgs.writeText "emulation-${name}" text;
-    })
-    profiles;
-
   coreOptionFiles = lib.mapAttrsToList
-    (name: text: {
-      inherit name;
-      file = pkgs.writeText "emulation-${name}" text;
+    (corename: text: {
+      inherit corename;
+      file = pkgs.writeText "emulation-${corename}.opt" text;
     })
     coreOptions;
 
@@ -332,21 +264,18 @@ let
     install -d -m 0755 -o ${cfg.user} -g ${cfg.group} \
       "${cfg.configRoot}/retroarch" \
       "${cfg.configRoot}/retroarch/autoconfig" \
-      "${cfg.configRoot}/retroarch/core-options" \
       "${cfg.configRoot}/retroarch/info" \
-      "${cfg.configRoot}/retroarch/profiles" \
       "${cfg.configRoot}/retroarch/remaps" \
       "${cfg.configRoot}/retroarch/shaders" \
       "${cfg.configRoot}/retroarch/shaders-user" \
-      "${cfg.configRoot}/retroarch/system-overrides" \
       "${cfg.dataRoot}/xdg/config/retroarch/config" \
       "${cfg.dataRoot}/logs/retroarch"
 
     install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${retroarchCfg} "${cfg.configRoot}/retroarch/retroarch.cfg"
     install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${globalShaderPreset} "${cfg.dataRoot}/xdg/config/retroarch/config/global.slangp"
-    install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${shaderPolicy} "${cfg.configRoot}/retroarch/shader-policy.json"
-    if [ -d "${packages.retroarchPackage}/share/libretro/info" ]; then
-      cp -R --no-preserve=mode,ownership "${packages.retroarchPackage}/share/libretro/info/." "${cfg.configRoot}/retroarch/info/" || true
+    rm -f "${cfg.configRoot}/retroarch/shader-policy.json"
+    if [ -d "${packages.libretroCoreInfo}/share/retroarch/cores" ]; then
+      cp -R --no-preserve=mode,ownership "${packages.libretroCoreInfo}/share/retroarch/cores/." "${cfg.configRoot}/retroarch/info/" || true
     fi
     chown -R ${cfg.user}:${cfg.group} "${cfg.configRoot}/retroarch/info" "${cfg.dataRoot}/logs/retroarch"
     touch "${cfg.dataRoot}/logs/retroarch/retroarch.log"
@@ -356,37 +285,20 @@ let
       cp -R --no-preserve=mode,ownership "${packages.joypadAutoconfig}/share/libretro/autoconfig/." "${cfg.configRoot}/retroarch/autoconfig/" || true
     fi
     install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${autoconfig8BitDo} "${cfg.configRoot}/retroarch/autoconfig/8BitDo Ultimate 2C Wireless Controller.cfg"
+    install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${autoconfig8BitDo} "${cfg.configRoot}/retroarch/autoconfig/udev/8BitDo Ultimate 2C Wireless Controller.cfg"
+    install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${autoconfig8BitDo} "${cfg.configRoot}/retroarch/autoconfig/udev/8BitDo_Ultimate_2C_Wireless_Controller_BT.cfg"
+    install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${autoconfig8BitDo} "${cfg.configRoot}/retroarch/autoconfig/udev/8BitDo_Ultimate_2C_Wireless_Controller_USB.cfg"
     install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${autoconfigSwitchPro} "${cfg.configRoot}/retroarch/autoconfig/Nintendo Switch Pro Controller.cfg"
     install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${autoconfigSwitchPro} "${cfg.configRoot}/retroarch/autoconfig/udev/Nintendo Switch Pro Controller.cfg"
 
-    install -m 0644 -o ${cfg.user} -g ${cfg.group} /dev/null "${cfg.configRoot}/retroarch/core-options/default.opt"
+    rm -rf "${cfg.configRoot}/retroarch/core-options"
+    rm -rf "${cfg.configRoot}/retroarch/profiles" "${cfg.configRoot}/retroarch/system-overrides"
+    find "${cfg.dataRoot}/xdg/config/retroarch/config" -mindepth 1 -maxdepth 2 -type f ! -name '*.opt' ! -name 'global.slangp' -delete
+    rm -f "${cfg.dataRoot}/xdg/config/retroarch/config/ParaLLEl N64/ParaLLEl N64.opt"
     ${lib.concatMapStringsSep "\n" (entry: ''
-      install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${entry.file} "${cfg.configRoot}/retroarch/core-options/${entry.name}"
+      install -d -m 0755 -o ${cfg.user} -g ${cfg.group} "${cfg.dataRoot}/xdg/config/retroarch/config/${entry.corename}"
+      install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${entry.file} "${cfg.dataRoot}/xdg/config/retroarch/config/${entry.corename}/${entry.corename}.opt"
     '') coreOptionFiles}
-    ${lib.concatMapStringsSep "\n" (entry: ''
-      install -D -m 0644 -o ${cfg.user} -g ${cfg.group} ${entry.file} "${cfg.configRoot}/retroarch/profiles/${entry.name}"
-    '') profileFiles}
-
-    printf '%s' '${config.ghostship.emulation.internal.lib.allSystemsJson}' | jq -c '.[]' | while read -r system; do
-      id="$(jq -r '.id' <<<"$system")"
-      emulator="$(jq -r '.emulator' <<<"$system")"
-      override="${cfg.configRoot}/retroarch/system-overrides/$id.cfg"
-      if [ -e "$override" ]; then
-        rm "$override"
-      fi
-    done
-
-    if [ ! -e "${cfg.configRoot}/retroarch/profiles/current.cfg" ]; then
-      ln -s "nnedi3-clean.cfg" "${cfg.configRoot}/retroarch/profiles/current.cfg"
-    else
-      current_target="$(readlink "${cfg.configRoot}/retroarch/profiles/current.cfg" 2>/dev/null || true)"
-      case "$current_target" in
-        ""|megabezel-auto.cfg)
-          ln -sfn "nnedi3-clean.cfg" "${cfg.configRoot}/retroarch/profiles/current.cfg"
-          ;;
-      esac
-    fi
-    chown -h ${cfg.user}:${cfg.group} "${cfg.configRoot}/retroarch/profiles/current.cfg" || true
 
     ln -sfn ${packages.shaderSlang}/share/libretro/shaders_slang "${cfg.configRoot}/retroarch/shaders/shaders_slang"
     ln -sfn ${packages.shaderGlsl}/share/libretro/shaders_glsl "${cfg.configRoot}/retroarch/shaders/shaders_glsl"
@@ -397,19 +309,15 @@ let
     set -euo pipefail
     export PATH=${config.ghostship.emulation.internal.lib.scriptPath}:$PATH
     shader_root="${cfg.configRoot}/retroarch/shaders"
-    status="${cfg.configRoot}/retroarch/shader-status.json"
+    status="${cfg.dataRoot}/logs/retroarch/shader-status.json"
     missing=0
     slang=false
     nnedi3=false
-    sharp=false
-    megabezel=false
     glsl=false
     cg=false
 
     if [ -e "$shader_root/shaders_slang" ]; then slang=true; else missing=1; fi
     if [ -e "$shader_root/shaders_slang/edge-smoothing/nnedi3/nnedi3-nns32-4x-rgb.slangp" ]; then nnedi3=true; else missing=1; fi
-    if [ -e "$shader_root/shaders_slang/pixel-art-scaling/sharp-bilinear-2x-prescale.slangp" ]; then sharp=true; else missing=1; fi
-    if [ -e "$shader_root/shaders_slang/bezel/Mega_Bezel" ]; then megabezel=true; else missing=1; fi
     if [ -e "$shader_root/shaders_glsl" ]; then glsl=true; else missing=1; fi
     if [ -e "$shader_root/shaders_cg" ]; then cg=true; else missing=1; fi
 
@@ -417,12 +325,10 @@ let
     jq -n \
       --argjson slang "$slang" \
       --argjson nnedi3 "$nnedi3" \
-      --argjson sharp "$sharp" \
-      --argjson megabezel "$megabezel" \
       --argjson glsl "$glsl" \
       --argjson cg "$cg" \
       --arg checked_at "$(date -u +%FT%TZ)" \
-      '{checked_at:$checked_at, slang:$slang, nnedi3:$nnedi3, sharp:$sharp, megabezel:$megabezel, glsl:$glsl, cg:$cg}' >"$status.tmp"
+      '{checked_at:$checked_at, slang:$slang, nnedi3:$nnedi3, glsl:$glsl, cg:$cg}' >"$status.tmp"
     chown ${cfg.user}:${cfg.group} "$status.tmp"
     chmod 0644 "$status.tmp"
     mv "$status.tmp" "$status"